@@ -1,6 +1,6 @@
 module github.com/fiatjaf/relayer
 
-go 1.15
+go 1.18
 
 require (
 	github.com/fiatjaf/go-nostr v0.7.3
@@ -11,4 +11,12 @@ require (
 	github.com/lib/pq v1.8.0
 	github.com/rs/cors v1.7.0
 	github.com/rs/zerolog v1.20.0
+)
+
+require (
+	github.com/btcsuite/btcd/btcec/v2 v2.2.0 // indirect
+	github.com/btcsuite/btcd/chaincfg/chainhash v1.0.1 // indirect
+	github.com/decred/dcrd/crypto/blake256 v1.0.0 // indirect
+	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.0.1 // indirect
+	github.com/valyala/fastjson v1.6.3 // indirect
 )
